backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.4
psycopg[binary]==3.2.9
pgvector==0.3.6
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
eval-type-backport==0.3.0
email-validator==2.2.0
python-jose[cryptography]==3.4.0
requests==2.32.4
boto3==1.40.12
reportlab==4.4.3
pillow==11.3.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
python-multipart==0.0.20
httpx==0.28.1
pypdf==5.9.0
pydicom==3.0.1
pytest==8.4.1
pytest-cov==6.2.1
ruff==0.12.9
